+++ src/services/signup/postRegister.ts
@@ -13,7 +13,7 @@ export interface PostRegisterProps {
 
 export const postRegister = async (formData: PostRegisterProps) => {
   try {
-    const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/auth/register`, {
+    const res = await fetch(`${apiUrl}/api/v1/auth/register`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
