+++ src/services/signup/postVerifyRequest.ts
@@ -16,17 +16,21 @@ export const postVerifyRequest = async (email: string) => {
 
     if (res.status === 500) {
       showToast("warning", "이미 가입된 계정입니다.", "");
+      throw new Error("이미 가입된 계정입니다.");
     }
 
     if (contentType && contentType.includes("application/json")) {
       const data = await res.json();
+      console.log("JSON response:", data);
+      
       if (!res.ok || data.success === false) {
         throw new Error(data.message || "인증번호 요청 실패");
       }
       return data;
     } else {
       const text = await res.text();
-      if (!res.ok || text !== "인증번호 요청이 완료되었습니다.") {
+      
+      if (!res.ok) {
         throw new Error(text || "인증번호 요청 실패");
       }
       return { success: true, message: text };
